End of phase 2. Unit tests done for User, Services, Orders and Suggestions services.

+++ src/main/java/com/maktabsharif/homeservices/domain/Suggestions.java
@@ -29,7 +29,7 @@ public class Suggestions{
     @Column(name = "start_date_by_expert")
     private Timestamp startDateByExpert;
     @Column(name = "order_do_duration")
-    private Long orderDoDuration;
+    private Integer orderDoDuration;
     @Column(name = "selected")
     private Boolean selecetd;
